pyproject.toml
[project]
name = "fake-news-gat"
version = "0.1.0"
description = "Fake News Propagation Detection using Graph Attention Networks"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    # Core ML/DL
    "torch>=2.1.0",
    "torch-geometric>=2.5.0",
    "transformers>=4.35.0",
    "sentence-transformers>=2.2.2",
    
    # Graph and Network Analysis
    "networkx>=3.2",
    "igraph>=0.11.0",
    
    # Data Processing
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    
    # NLP and Text Processing
    "nltk>=3.8.1",
    "spacy>=3.7.0",
    "regex>=2023.10.0",
    
    # Machine Learning
    "scikit-learn>=1.3.0",
    
    # Visualization
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "plotly>=5.18.0",
    
    # Configuration and Utilities
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
    "click>=8.1.0",
    
    # Experiment Tracking
    "tensorboard>=2.15.0",
    "wandb>=0.16.0",
    
    # Data Validation
    "pydantic>=2.5.0",
    
    # Jupyter
    "jupyter>=1.0.0",
    "ipykernel>=6.27.0",
    "ipywidgets>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=html --cov-report=term-missing"
